Regularly update OpenID configuration
Required some refactoring
