+++ Cargo.toml
@@ -10,9 +10,8 @@ jsonwebtoken = "8"
 serde = {version = "1", features = ["derive"] }
 serde_json = "1"
 lazy_static="1"
-reqwest = { version = "0.10", features = ["blocking", "json"] }
+reqwest = { version = "0.11", features = ["json"] }
 serde_derive = "1"
-once_cell = "1.8"
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-tokio = { version = "1.0", features = ["full"] }
+tokio = { version = "1", features = ["full"] }
